deleted Smurfs.js, unsure how to make props work between Smurf.js and Smurfs.js with context api, kept saying undefined with mapping through <Smurfs />, removed component and is now rendering fine with context api, :shrug:
